Update README.md
Using partially filled arrays with a sentinel value

Step 1

Examine and run the code to see how it functions. Add a comment to the code each time the grades array is used, describing what that line of code is doing. You should have at least 6 comments added. Why is the use of this array called a "partially filled array with a sentinel value"? What is the value of the sentinel?

Step 2

Add statements to also calculate the average of the elements inputted and display the average to the screen.

Step 3

Also determine the highest grade and the lowest grade entered.